requirements.txt
sqlalchemy
alembic
psycopg2-binary
python-dotenv
fastapi
uvicorn
aiogram
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-multipart
httpx
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
